better organise types for get functions
